requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

requests==2.32.3
beautifulsoup4==4.12.3

pydantic==2.6.4

langchain==0.3.25
langchain-google-genai==2.1.4
langchain-mistral-ai

python-multipart==0.0.9
httpx==0.27.0

python-dotenv==1.0.1
